moved rag_manager to os_assistant lib
