+++ src/os_assistant/rag_manager.py
